:beer: adding map

+++ src/main/java/com/co/meli/coupon/delegate/ItemPriceDelegate.java
@@ -1,6 +1,6 @@
 package com.co.meli.coupon.delegate;
 
 public interface ItemPriceDelegate {
-    public Double getItemPriceItem(String idItem);
+    public Float getItemPriceItem(String idItem);
 
 }
